Separate out boost control from auxiliaries.*
