+++ speeduino/board_teensy35.cpp
@@ -16,6 +16,7 @@
 #include "scheduler_ignition_controller.h"
 #include "scheduler_fuel_controller.h"
 #include "src/controllers/fan/fanController.h"
+#include "src/controllers/boost/boostController.h"
 
  //These are declared locally in comms_CAN now due to this issue: https://github.com/tonton81/FlexCAN_T4/issues/67
 // #if defined(__MK64FX512__)         // use for Teensy 3.5 only
