+++ speeduino/src/controllers/boost/boostController.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <stdint.h>
+
+void initialiseBoost(uint8_t boostPin);
+void boostControl(void);
+void boostDisable(void);
+
+void boostInterrupt(void);
+
